Refactoring: class GeoDataPointmap::BreakLineAddCommand moved to independent files.
